Refactorización N: 10 Tipo de refactorización: Move Method Clases modificadas: ListaMenuDesplegable y ListaCelulares
