+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/AgilProject2.iml" filepath="$PROJECT_DIR$/AgilProject2.iml" />
+      <module fileurl="file://$PROJECT_DIR$/AgilProject1.iml" filepath="$PROJECT_DIR$/AgilProject1.iml" />
     </modules>
   </component>
 </project>
